requirements.txt
# TextTorch - Requirements
# PyTorch NLP Project for Text Classification

# Deep Learning
# PyTorch CPU-only (recomendado para Streamlit Cloud - economiza RAM e espaço)
# Para instalar localmente: pip install torch torchvision --index-url https://download.pytorch.org/whl/cpu
torch>=2.0.0
torchvision>=0.15.0

# NLP & Machine Learning
scikit-learn>=1.3.0
nltk>=3.8.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=14.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
joblib>=1.3.0
pyyaml>=6.0
tqdm>=4.65.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0
notebook>=7.0.0

# Additional utilities
python-dotenv>=1.0.0

# Streamlit
streamlit>=1.28.0
